% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.bbcor}
\alias{plot.bbcor}
\title{Plot bbcor point estimates and intervals}
\usage{
\method{plot}{bbcor}(x, ci = 0.9, point_col = "red", bar_col = "black", ...)
}
\arguments{
\item{x}{An object of class \code{bbcor}}

\item{ci}{Width of credible interval. Defaults to 0.9.}

\item{point_col}{Color for point indicating mean of posterior}

\item{bar_col}{Color of bar for credible interval}

\item{...}{Currently ignored}
}
\value{
An object of class \code{ggplot}
}
\description{
Plot bbcor point estimates and intervals
}
\examples{
Y <- BGGM::ptsd
bb <- bbcor(Y)
plot(bb)

}
