% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{print.bayeslincom}
\alias{print.bayeslincom}
\title{Print formatted summary of a \code{bayeslincom} object}
\usage{
\method{print}{bayeslincom}(x, decimals = 2, ...)
}
\arguments{
\item{x}{An object of class \code{bayeslincom}}

\item{decimals}{The number of decimals points to which estimates should be rounded}

\item{...}{Other arguments to be passed to \code{print}}
}
\value{
A formatted summary of posterior samples
}
\description{
Print formatted summary of a \code{bayeslincom} object
}
