% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbcor.R
\name{print.bbcor}
\alias{print.bbcor}
\title{Print \code{bbcor} Objects}
\usage{
\method{print}{bbcor}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bbcor}}

\item{...}{Currently ignored}
}
\description{
Print the correlation or partial correlation matrix
}
