% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_samples.R
\name{posterior_samples}
\alias{posterior_samples}
\title{Extract Posterior Samples}
\usage{
posterior_samples(object, ...)
}
\arguments{
\item{object}{An object of class \code{bbcor}}

\item{...}{Currently ignored}
}
\value{
A data frame including the posterior samples
}
\description{
Extract Posterior Samples
}
\examples{

Y <- mtcars[,1:5]

bb_samps <- bbcor(Y, method = "spearman")

# correlations
posterior_samples(bb_samps)


# partial correlations
posterior_samples(cor_2_pcor(bb_samps))
}
